+++ lib/pkg.ts
@@ -1,4 +1,4 @@
 export default {
-  version: '1.3.1-beta.3',
+  version: '1.3.1-beta.5',
   name: 'aliyun-pds-js-sdk',
 }